Move ls to client.py
The ls function is useful in other contexts than just the ls program.
The program itself depends on functions in .client [1], which would quickly
lead to import cycles. Simply moving the function to client.py resolves
this.

[1] client.py should probably be split up at some point
