+++ python/oneseismic/client/client.py
@@ -330,3 +330,25 @@ class http_session(requests.Session):
         cfg = config(cache_dir = cache_dir).load()
         auth = tokens(cache_dir = cache_dir).load(cfg)
         return http_session(base_url = cfg['url'], tokens = auth)
+
+def ls(session):
+    """List available cubes
+
+    List the cubes stored in oneseismic. The ids returned should all be valid
+    arguments for the oneseismic.client.cube class.
+
+    Parameters
+    ----------
+    session : oneseismic.http_session
+        Session with authorization headers set
+
+    Returns
+    -------
+    guids : iterable of str
+        Cube GUIDs
+
+    See also
+    --------
+    oneseismic.client.cube
+    """
+    return session.get('query').json()['links'].keys()
